# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Diamond Sealing — subcontractor operations management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /monthly-rankings:
    get:
      operationId: listMonthlyRankings
      tags: [awards]
      parameters:
        - name: month
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Monthly rankings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MonthlyRanking" }
  /monthly-rankings/calculate:
    post:
      operationId: calculateMonthlyRankings
      tags: [awards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [month]
              properties:
                month: { type: string }
      responses:
        "200":
          description: Calculated rankings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MonthlyRanking" }
  /scoring-weights:
    get:
      operationId: getScoringWeights
      tags: [awards]
      responses:
        "200":
          description: Current scoring weights
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScoringWeights" }
    patch:
      operationId: updateScoringWeights
      tags: [awards]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateScoringWeightsBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScoringWeights" }
  /monthly-awards:
    get:
      operationId: listMonthlyAwards
      tags: [awards]
      responses:
        "200":
          description: Monthly awards
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MonthlyAward" }
    post:
      operationId: createMonthlyAward
      tags: [awards]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMonthlyAwardBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonthlyAward" }
  /monthly-awards/{id}:
    patch:
      operationId: updateMonthlyAward
      tags: [awards]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMonthlyAwardBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonthlyAward" }
...
components:
  schemas:
...
    MonthlyRanking:
      type: object
      required: [id, subcontractorId, month, totalScore]
      properties:
        id: { type: integer }
        subcontractorId: { type: integer }
        subcontractorName: { type: string }
        month: { type: string }
        rank: { type: ["integer", "null"] }
        totalScore: { type: number }
        metresScore: { type: number }
        metresPerHourScore: { type: number }
        auditScore: { type: number }
        punctualityScore: { type: number }
        photoComplianceScore: { type: number }
        callbackScore: { type: number }
        attendanceScore: { type: number }
        totalMetres: { type: number }
        avgMetresPerHour: { type: number }
        daysWorked: { type: integer }
        jobsCompleted: { type: integer }
        callbackCount: { type: integer }
        lateArrivals: { type: integer }
        missingPhotoJobs: { type: integer }
        auditFlagCount: { type: integer }
    ScoringWeights:
      type: object
      properties:
        id: { type: integer }
        metresWeight: { type: number }
        metresPerHourWeight: { type: number }
        auditWeight: { type: number }
        punctualityWeight: { type: number }
        photoComplianceWeight: { type: number }
        callbackWeight: { type: number }
        attendanceWeight: { type: number }
    UpdateScoringWeightsBody:
      type: object
      properties:
        metresWeight: { type: number }
        metresPerHourWeight: { type: number }
        auditWeight: { type: number }
        punctualityWeight: { type: number }
        photoComplianceWeight: { type: number }
        callbackWeight: { type: number }
        attendanceWeight: { type: number }
    MonthlyAward:
      type: object
      required: [id, month, winnerId, awardType, awardTitle, reasonText]
      properties:
        id: { type: integer }
        month: { type: string }
        winnerId: { type: integer }
        winnerName: { type: string }
        winnerPhoto: { type: ["string", "null"] }
        awardType: { type: string, enum: [weekend_away, tv, experience, voucher, cash, custom] }
        awardTitle: { type: string }
        awardDescription: { type: ["string", "null"] }
        awardValue: { type: ["number", "null"] }
        reasonText: { type: string }
        totalScore: { type: ["number", "null"] }
        adminApproved: { type: boolean }
        publishedToStaff: { type: boolean }
        publishedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    CreateMonthlyAwardBody:
      type: object
      required: [month, winnerId, awardType, awardTitle, reasonText]
      properties:
        month: { type: string }
        winnerId: { type: integer }
        awardType: { type: string, enum: [weekend_away, tv, experience, voucher, cash, custom] }
        awardTitle: { type: string }
        awardDescription: { type: string }
        awardValue: { type: number }
        winnerPhoto: { type: string }
        reasonText: { type: string }
        totalScore: { type: number }
    UpdateMonthlyAwardBody:
      type: object
      properties:
        awardType: { type: string, enum: [weekend_away, tv, experience, voucher, cash, custom] }
        awardTitle: { type: string }
        awardDescription: { type: string }
        awardValue: { type: number }
        winnerPhoto: { type: string }
        reasonText: { type: string }
        adminApproved: { type: boolean }
        publishedToStaff: { type: boolean }

    # ─── Worker Profiles & Skills ──────────────────────────────────────────
...